fix vscode changes not triggering ~
when saving files, IDEA changes the modification date of the entire folder. vscode changes it for the file only. technically, only watching the folder modification date is more performant, but since it wasn't working with the editor I prefer to use for elm I'm doing this and I don't care